Update probe-javassist to Java 11 and extract MethodInstrumenter
