+++ monitoring-probe-javassist/build.gradle
@@ -5,8 +5,8 @@ plugins {
 	id 'maven-publish'
 }
 
-sourceCompatibility = 8
-targetCompatibility = 8
+sourceCompatibility = 11
+targetCompatibility = 11
 
 // The following lines are necessay to create a test jar
 project.configurations {
